🧪  dissolve store helpers
